fixed adding and subtracting order total

+++ db/models/product.js
@@ -6,7 +6,7 @@ const db = require('APP/db')
 const Product = db.define('products', {
 	name: Sequelize.STRING,
 	imageURL: Sequelize.STRING,
-	price: Sequelize.DECIMAL(10, 2),
+	price: Sequelize.INTEGER,
 	description: Sequelize.TEXT,
 	inventory: Sequelize.INTEGER,
 	tags: Sequelize.ARRAY(Sequelize.STRING)
